feat: Add program output manager
Reason for change:
Previous program output design has logic spread throughout the
LDAR-Sim object. This made it more difficult to isolate
the logic responsible for program outputs.

Resolution:
Moved all output logic for program outputs to a new class.
This centralizes the logic for program outputs and makes it more
accessible.

Effect(s) of change:
N/A
